use off_t for file positions to avoid type warnings
git-svn-id: svn+ssh://frodo/cluster/svn/projects/slash_nara@16036 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda

+++ mount_slash/mount_slash.h
@@ -97,7 +97,7 @@ PSCTHR_MKCAST(msfsthr, msfs_thread, MSTHRT_FS);
 struct msl_ra {
 	off_t				 mra_loff;	/* last offset */
 	off_t				 mra_raoff;	/* current read ahead offset */
-	size_t				 mra_lsz;	/* last size */
+	off_t				 mra_lsz;	/* last size */
 	int				 mra_nseq;	/* num sequential io's */
 	int				 mra_nrios;	/* num read io's */
 	int				 mra_bkwd;	/* reverse access io */
